refactor: replace executor with execution client
